move common functions to dedicated fie
