+++ tests/integration/reconcile/test_aggregates_recon_capture.py
@@ -14,7 +14,7 @@ from databricks.labs.remorph.reconcile.recon_output_config import (
     AggregateQueryOutput,
 )
 from databricks.labs.remorph.reconcile.recon_capture import generate_final_reconcile_aggregate_output
-from tests.integration.reconcile.test_aggregates_reconcile import expected_reconcile_output_dict, expected_rule_output
+from .common import expected_reconcile_output_dict, expected_rule_output
 
 
 def remove_directory_recursively(directory_path):
